Added w=majority to URI
If you're out of the loop, Heroku doesn't work with commands containing `&`, therefore the URI will cause an error. the workaround is to just manually place w=majority (The usefullness of the option is debatable but it's the default). In the previous commit, this has been moved out of the database loader